+++ mz_zip.h
@@ -191,7 +191,7 @@ int32_t mz_zip_extrafield_find(void *stream, uint16_t type, uint16_t *length);
 int32_t mz_zip_extrafield_read(void *stream, uint16_t *type, uint16_t *length);
 // Reads an extrafield header from a stream
 
-int32_t mz_zip_extrafield_write(void *stream, uint32_t type, uint16_t length);
+int32_t mz_zip_extrafield_write(void *stream, uint16_t type, uint16_t length);
 // Writes an extrafield header to a stream
 
 /***************************************************************************/
